% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/be-api-key.R
\name{be_auth_token}
\alias{be_auth_token}
\title{Get or set BINARYEDGE_API_KEY value}
\usage{
be_auth_token(force = FALSE)
}
\arguments{
\item{force}{force setting a new omdb API key for the current environment?}
}
\value{
atomic character vector containing the BinaryEdge API key
}
\description{
The API wrapper functions in this package all rely on a BinaryEdge API
key residing in the environment variable \code{BINARYEDGE_API_KEY}. The
easiest way to accomplish this is to set it in the `~.Renviron`` file in your
home directory.
}
