% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryedge-package.R
\docType{package}
\name{binaryedge}
\alias{binaryedge}
\alias{binaryedge-package}
\title{Tools to Query and Orchestrate the 'BinaryEdge' 'API'}
\description{
'BinaryEdge' acquires, analyzes and classifies internet-wide data
by combining efforts in the areas of 'cybersecurity', engineering and data science.
Tools are provided to work with the 'BinaryEdge' 'API' <app.binaryedge.io>.
}
\details{
\itemize{
\item URL: \url{https://github.com/balgan/rbinaryedge}
\item BugReports: \url{https://github.com/balgan/rbinaryedge/issues}
}

\if{html}{
\figure{be-logo.png}{options: width="250" alt="Figure: be-logo.png"}
}

\if{latex}{
\figure{be-logo.png}{options: width=10cm}
}
}
\note{
To use this package you will need to go to \url{https://app.binaryedge.io}
and register for an account.
}
\author{
Balgan \href{mailto:balgan@binaryedge.io}{balgan@binaryedge.io}
}
