% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnsipQuery.R
\name{dnsipQuery}
\alias{dnsipQuery}
\title{Query DNS database - IP}
\usage{
dnsipQuery(ip, page, auth_token)
}
\arguments{
\item{ip}{target ip}

\item{page}{page number of query}

\item{auth_token}{authentication token for the API}
}
\description{
Query dns database to get list dns information for a IP
}
