% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domaindataleakQuery.R
\name{domaindataleakQuery}
\alias{domaindataleakQuery}
\title{Query dataleak database - domain}
\usage{
domaindataleakQuery(domain, auth_token)
}
\arguments{
\item{domain}{target domain}

\item{auth_token}{authentication token for the API}
}
\description{
Query dataleak database for domain
}
