% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historicalQuery.R
\name{historicalQuery}
\alias{historicalQuery}
\title{Query host database}
\usage{
historicalQuery(ip, auth_token)
}
\arguments{
\item{ip}{target IP address}

\item{auth_token}{authentication token for the API}
}
\description{
Query historical endpoint to extract information on CIDR or IP
}
