% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageQuery.R
\name{imageQuery}
\alias{imageQuery}
\title{Query image IP}
\usage{
imageQuery(ip, page, auth_token)
}
\arguments{
\item{ip}{target ip address}

\item{page}{page number of query}

\item{auth_token}{authentication token for the API}
}
\description{
Query to get details of an image in an IP address
}
