% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskscoreIP.R
\name{riskscoreQuery}
\alias{riskscoreQuery}
\title{Query Score for IP}
\usage{
riskscoreQuery(ip, auth_token)
}
\arguments{
\item{ip}{target IP address}

\item{auth_token}{authentication token for the API}
}
\description{
Query risk score for ip address
}
