% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchQuery.R
\name{searchQuery}
\alias{searchQuery}
\title{Query search}
\usage{
searchQuery(searchparameter, page, auth_token)
}
\arguments{
\item{searchparameter}{search parameter to use on query}

\item{page}{page number of query}

\item{auth_token}{authentication token for the API}
}
\description{
Query for search to look for products or specific countries
}
