% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torrentQuery.R
\name{torrentQuery}
\alias{torrentQuery}
\title{Query torrent database}
\usage{
torrentQuery(ip, auth_token)
}
\arguments{
\item{ip}{target IP address}

\item{auth_token}{authentication token for the API}
}
\description{
Query torrent database to extract information on CIDR or IP that downloaded torrents
}
