% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaildataleakQuery.R
\name{emaildataleakQuery}
\alias{emaildataleakQuery}
\title{Query dataleak database - email}
\usage{
emaildataleakQuery(email, auth_token)
}
\arguments{
\item{email}{target email address}

\item{auth_token}{authentication token for the API}
}
\description{
Query dataleak database for single email
}
