% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historicaltorrentQuery.R
\name{historicaltorrentQuery}
\alias{historicaltorrentQuery}
\title{Query historical torrent}
\usage{
historicaltorrentQuery(ip, auth_token)
}
\arguments{
\item{ip}{target IP address}

\item{auth_token}{authentication token for the API}
}
\description{
Query historical torrent  endpoint to extract information on IP
}
